refactor: remove dead evaluator and syscall adapter helpers
- delete NewStandardEvaluator and migrate tests to NewStandardEvaluatorWithStores

- remove dead syscall store adapter implementation and dedicated adapter tests

- move shared syscall store mock into network analyzer tests

- record Phase B decisions and completion status in task 0122 plan
